update all and remove logic
